requirements.txt
flask>=3.0.0
pillow>=10.0.0
qrcode[pil]>=7.4.2
werkzeug>=3.0.0
pyngrok>=6.0.0